Move generic components to fields dir
